fix: $pc change and relative linking for pseudo instruction

+++ src/rarsreborn/core/compilation/decoder/DecodingResult.java
@@ -2,4 +2,4 @@ package rarsreborn.core.compilation.decoder;
 
 import rarsreborn.core.core.instruction.IInstruction;
 
-public record DecodingResult(IInstruction instruction, int bytesConsumed) {}
+public record DecodingResult(IInstruction instruction, byte bytesConsumed) {}
